+++ src/views/ResetPassword/ResetPassword.scss
@@ -17,7 +17,7 @@
         height: 500px;
         border-radius: 10px;
         background-color: white;
-        box-shadow: 2.5px 2.5px 2px #88888881;
+        box-shadow: 1px 2px 8px -1px rgba(0,0,0,0.56);
         display: flex;
         flex-direction: column;
         align-items: center;
@@ -25,11 +25,6 @@
         form {
             margin-top: 50%;
             transform: translateY(-50%);
-
-            button {
-                width: auto;
-                margin: auto;
-            }
         }
     }
 }
